requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
pyTelegramBotAPI==4.24.0
python-dotenv==1.0.1
requests==2.32.3
PyPDF2==3.0.1
python-docx==1.1.2
python-pptx==1.0.2
pytest==8.3.4
pytest-asyncio==0.24.0
openai==1.57.4
pydantic==2.10.3
python-multipart==0.0.20
